+++ DataTransferApp.Net/Views/ChangesWindow.xaml.cs
@@ -1,4 +1,6 @@
+using System.Diagnostics;
 using System.Windows;
+using System.Windows.Input;
 using DataTransferApp.Net.ViewModels;
 
 namespace DataTransferApp.Net.Views
@@ -17,6 +19,14 @@ namespace DataTransferApp.Net.Views
         private void CloseButton_Click(object sender, RoutedEventArgs e)
         {
             this.Close();
+        }
+
+        private void OpenHyperlink(object sender, ExecutedRoutedEventArgs e)
+        {
+            if (e.Parameter is string url && !string.IsNullOrEmpty(url))
+            {
+                Process.Start(new ProcessStartInfo(url) { UseShellExecute = true });
+            }
         }
     }
 }